requirements.txt
Django==2.2.4
django-debug-toolbar==1.9.1
django-nose==1.4.6
django-tinymce4-lite==1.7.5
Jinja2==2.10
jsmin==2.2.2
MarkupSafe==1.0
Pillow==5.1.0
python-slugify==3.0.3
pytz==2018.5
sqlparse==0.2.4
text-unidecode==1.2
